requirements.txt
urlextract
pandas
python-Levenshtein
supabase
python-dotenv
tldextract
